MALLM_backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
google-api-python-client
python-dotenv
beautifulsoup4
google-generativeai
cohere
streamlit